core: moved to core/swapper
