+++ core/swapper/swapper_suite_test.go
@@ -1,1 +1,13 @@
 package swapper_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSwapper(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Swapper Suite")
+}
