--- core/swapper/swapper_test.go
@@ -1,1 +1,0 @@
-package swapper_test
